practice/requirements.txt
mcp[cli]==1.8.0 
mcp
google-generativeai==0.8.5
python-dotenv
ipykernel
httpx

